requirements.txt
discord~=1.0.1
mysql-connector-python~=8.0.26
requests~=2.25.1